Move endpoints into their own folder
