+++ src/endpoints/translate.js
@@ -1,7 +1,7 @@
 const fetch = require('node-fetch').default;
 const https = require('https');
 const { readSecret, SECRET_KEYS } = require('./secrets');
-const { getConfigValue } = require('./util');
+const { getConfigValue } = require('../util');
 
 const DEEPLX_URL_DEFAULT = 'http://127.0.0.1:1188/translate';
 const ONERING_URL_DEFAULT = 'http://127.0.0.1:4990/translate';
